fixed layers removal bug and cleaned up some code
