+++ backend/common/ai_drive.py
@@ -232,16 +232,7 @@ def ml_drive(
         else:
             if json_csv_data_str:
                 input_df = pd.read_json(json_csv_data_str, orient="records")
-                count = -1
-                valsToNums = {}
-                def to_nums(val):
-                    if val in valsToNums:
-                        return float(valsToNums[val])
-                    nonlocal count
-                    count += 1
-                    valsToNums[val] = count
-                    return float(count)
-                input_df[target] = input_df[target].map(to_nums)
+                input_df[target] = input_df[target].astype("category").cat.codes
                 y = input_df[target]
                 X = input_df[features]
                 print(input_df.head())
