+++ frontend/playground-frontend/src/components/Home/CodeSnippet.js
@@ -4,8 +4,8 @@ import { FaCopy } from "react-icons/fa";
 import { toast } from "react-toastify";
 
 const CodeSnippet = (props) => {
-  const { backendResponse, layers } = props;
-
+  const { backendResponse } = props;
+  const layers = backendResponse?.auxiliary_outputs.user_arch;
   if (!backendResponse?.success) {
     return (
       backendResponse?.message || (
@@ -19,7 +19,7 @@ const CodeSnippet = (props) => {
       <textarea
         id="code-snippet-text"
         readOnly
-        rows="10"
+        rows={10}
         value={codeSnippetFormat(layers)}
       />
       <button
@@ -37,7 +37,7 @@ const CodeSnippet = (props) => {
 
 /**
  * This function returns necessary code skeleton to train data from local terminal
- * @param {layers[]} layers
+ * @param {string[]} layers
  * @returns string with correct python syntax to 'train' data
  */
 function codeSnippetFormat(layers) {
@@ -63,7 +63,7 @@ function codeSnippetFormat(layers) {
 /**
  * Given an array of layers, this function builds a string with all elements of the array after they have applied
  * the layerToString() function
- * @param {layers[]} layers
+ * @param {string[]} layers
  * @returns string in form of 'self.model = nn.Sequential(*[layerToString(layers[0]),... ,layerToString(layers[N-1])])
  */
 function layersToString(layers) {
@@ -71,53 +71,15 @@ function layersToString(layers) {
   let layersToString = prepend;
   const resultingList = [];
   for (let i = 0; i < layers.length; i++) {
-    resultingList.push(layerToString(layers[i]));
+    resultingList.push(layers[i]);
   }
   layersToString += resultingList.join(",") + "])";
   return layersToString;
 }
 
-/**
- * Depending on layer passed in, this function builds a string with layer's name, and parameters associated to it (if any)
- * @param {layers} layer
- * @returns string in form of <layer name>(<parameters>)
- */
-export function layerToString(layer) {
-  let layerToString = layer.object_name + "(";
-
-  if (layer.parameters !== undefined && layer.parameters !== null) {
-    const params = Object.keys(layer.parameters);
-    // params : [0: "inputSize", 1:"outputSize"]
-    if (params !== null && params !== undefined && params.length !== 0) {
-      // const paramList= Array{[params.length]}
-
-      const paramList = new Array(params.length);
-      for (let i = 0; i < params.length; i++) {
-        const param = params[i];
-        // param: "inputSize"
-
-        if (typeof layer.parameters[param] !== "undefined") {
-          paramList[layer.parameters[param].index] =
-            layer.parameters[param].value;
-        }
-      }
-      for (let i = 0; i < paramList.length; i++) {
-        layerToString += paramList[i];
-        layerToString += ",";
-      }
-
-      layerToString = layerToString.split("");
-      layerToString[layerToString.length - 1] = "";
-      layerToString = layerToString.join("");
-      // layerToString = layerToString.substring(0, layerToString.length)
-    }
-  }
-  layerToString += ")";
-  return layerToString;
-}
-
 CodeSnippet.propTypes = {
   backendResponse: PropTypes.shape({
+    auxiliary_outputs: PropTypes.object,
     success: PropTypes.bool,
     message: PropTypes.string,
   }),
